split mrealm module
